"Endpoint id" concept
The endpoint id is up to 10 bits of unique id extracted from the
datacenter's id by eliminating a prefix that is common to all datacenters
in the cluster.

This also introduces the idea of the "common prefix", which is the number
of bits of address that are common across all datacenters. As well as
being used to extract the endpoint id, this will also *later* be used to
identify which routes in the route table are subject to our management,
so we can safely delete unneeded routes from the table without affecting
other special routes like the default route.